[dataset] Add tf.dataset preprocessing for object_detection
- In order to make it independant from the classes the conversion to
one_hot vector of the labels has been moved to the sampling parts of the
fast_rcnn head.

- Add a dimension to the NUM_BOXES field:

During the fit iteration if the tensor as the shape (1,) tf.Dataset
automatically add a dimension => (1, ). To make the code works with it convert
the code to work with this extra dimension.

- Add doc for fast_rcnn and abstract_detection head