[pooling] Fix the shape issue whenever there was only one box to a pyramid level.
When there was one box the shape of the boxes was [4,] instead of [1, 4].
The crop_and_resize operations need a 2-D tensor in order to work.

Problem:

The issue came from this line:
levels = [tf.squeeze(tf.where(tf.equal(box_levels, i))) for i in range(num_level)]

It removed the proper dimension when there was only one box.