[serving] Add a serving arguments to abstract detection to bypass the check from tensorflow serving.
Will allow to bypass the save_model behavior the graph in serving mode.
Currently, the issue is that in training the ground_truths are passed to the call method but
not in inference. For the serving only the `images` and `images_information` are defined.
It means the inputs link to the ground_truths won't be defined in serving. However, in tensorflow
when the `training` arguments is defined int the method `call`, `tf.save_model.save` method
performs a check on the graph for training=False and training=True.
However, we don't want this check to be perform because our ground_truths inputs aren't defined.