:recycle: Add new resnet support and matcher target_assigner handle batch natively
1. target_assigner/matcher

Upgrade target_assigner and matcher to handle batch
and padding

- Matcher was inspired from detectron 2 (already inspired from tensorflow/models)
- Remove batch_assign_target. Target assigner can handle batch by itself.
- box_ops.compute_iou handle batch
- Since None for the batch_size can be handle in training saved_model do not
error anymore

2. Backbone

Add support for resnet with the pytorch style, which is used in
detectron2, tensorpack...

3. Rpn postprocessing

The postprocessing extract a fix set of anchors per layer of the
pyramid.

In the past the extraction was done before on the concatenation on the
predictions of all the layers. It may forces the network to take has many small
and big objects.